+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,16 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Andrey Zhelev on 10.12.2023.
+//
+
+import Foundation
+
+
+// структура данных для массива вопросов
+private struct QuizQuestion {
+    let image: String // название фильма / картинки
+    let text: String // вопрос по фильму
+    let correctAnswer: Bool // правильный ответ на вопрос Да / Нет
+}
